requirements.txt
Flask==3.1.0
tensorflow==2.19.0
python-dotenv==1.1.0
Pillow==11.2.1
imagehash==4.3.2
gunicorn==20.1.0
